fix: rename os to system
