+++ src/helpers/report.cjs
@@ -2,7 +2,7 @@ const { formatErrorAjv, validateReportAjv, validateReportConfigurationAjv } = re
 const { getContext, hasContext } = require('./github.cjs');
 const { readFileSync, writeFileSync } = require('node:fs');
 const { relative, sep: platformSeparator, resolve } = require('node:path');
-const { getOperatingSystem } = require('./os.cjs');
+const { getOperatingSystem } = require('./system.cjs');
 const { join } = require('node:path/posix');
 const { minimatch } = require('minimatch');
 const { randomUUID } = require('node:crypto');
